Refactor rkt into one file.
This will make it easy to add build labels and stub out in non-Linux
builds.
