+++ pkg/kubelet/rkt/pod_info.go
@@ -169,47 +169,3 @@ func splitLineByTab(line string) []string {
 	}
 	return result
 }
-
-// getPodInfos returns a map of [pod-uuid]:*podInfo
-func (r *Runtime) getPodInfos() (map[string]*podInfo, error) {
-	result := make(map[string]*podInfo)
-
-	output, err := r.runCommand("list", "--no-legend", "--full")
-	if err != nil {
-		return result, err
-	}
-	if len(output) == 0 {
-		// No pods are running.
-		return result, nil
-	}
-
-	// Example output of current 'rkt list --full' (version == 0.4.2):
-	// UUID                                 ACI     STATE      NETWORKS
-	// 2372bc17-47cb-43fb-8d78-20b31729feda	foo     running    default:ip4=172.16.28.3
-	//                                      bar
-	// 40e2813b-9d5d-4146-a817-0de92646da96 foo     exited
-	// 40e2813b-9d5d-4146-a817-0de92646da96 bar     exited
-	//
-	// With '--no-legend', the first line is eliminated.
-	for _, line := range output {
-		tuples := splitLineByTab(line)
-		if len(tuples) < 3 { // At least it should have 3 entries.
-			glog.Warningf("rkt: Unrecognized line: %q", line)
-			continue
-		}
-		id := tuples[0]
-
-		status, err := r.runCommand("status", id)
-		if err != nil {
-			glog.Errorf("rkt: Cannot get status for pod (uuid=%q): %v", id, err)
-			continue
-		}
-		info, err := parsePodInfo(status)
-		if err != nil {
-			glog.Errorf("rkt: Cannot parse status for pod (uuid=%q): %v", id, err)
-			continue
-		}
-		result[id] = info
-	}
-	return result, nil
-}
